+++ cli/cli/Services/VersionService.cs
@@ -1,3 +1,4 @@
+using Beamable.Common;
 using Beamable.Common.Util;
 using Newtonsoft.Json;
 using Serilog;
@@ -77,6 +78,29 @@ public class VersionService
 			info.installType = VersionInstallType.LocalTool;
 		}
 		return info;
+	}
+
+	public static PackageVersion GetExecutingCliVersion()
+	{
+		var versionString = BeamAssemblyVersionUtil.GetVersion<BeamoService>();
+		if (!PackageVersion.TryFromSemanticVersionString(versionString, out var packageVersion))
+		{
+			return "0.0.0";
+		}
+		return packageVersion;
+	}
+
+	public static PackageVersion GetNugetPackagesForExecutingCliVersion()
+	{
+		var currentVersion = GetExecutingCliVersion();
+		if (currentVersion.Major == 0)
+		{
+			// if the major is 0, then its likely 0.0.0 or 0.0.123, 
+			//  which means we want to use our local dev nuget package version, which is 0.0.123
+			return "0.0.123";
+		}
+
+		return currentVersion;
 	}
 
 	public struct VersionInfo