+++ client/Packages/com.beamable.server/Editor/Usam/Editor/LibrariesPaths.cs
@@ -1,0 +1,12 @@
+using Beamable.Editor.BeamCli.Commands;
+using System;
+using System.Collections.Generic;
+
+namespace Beamable.Server.Editor.Usam
+{
+	[Serializable]
+	public class LibrariesPaths
+	{
+		public List<BeamDependencyData> libraries = new List<BeamDependencyData>();
+	}
+}