+++ client/Packages/com.beamable/CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Cid and Pid are stored in `Beam.GlobalScope`
 - Dotnet is automatically installed to the project's /Library folder
 - Beam CLI is automatically installed to the project's /Library folder
+- Baked content exists in a separate memory location than the deserialized `content.json` CDN cache
 
 ### Fixed
 
@@ -35,6 +36,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Disabling Content inspectors no longer causes compiler errors.
 - Leaderboard `rankgt` field is not null when specifying outlier.
 - Fix renaming content throwing infinite warnings
+- Content cache evicts old content for new content id versions
+- `PlatformRequester` will reattempt failed requests caused by SSL connection issues.
 
 ### Removed